MAINT Move the pls to the cca to a cross_decomposition module
