+++ sklearn/__init__.py
@@ -69,7 +69,8 @@ else:
                'feature_selection', 'semi_supervised',
                'gaussian_process', 'grid_search', 'hmm', 'lda', 'linear_model',
                'metrics', 'mixture', 'naive_bayes', 'neighbors', 'pipeline',
-               'preprocessing', 'qda', 'svm', 'test', 'clone', 'pls',
+               'preprocessing', 'qda', 'svm', 'test', 'clone',
+               'cross_decomposition',
                'isotonic']
