+++ sklearn/cross_decomposition/__init__.py
@@ -1,0 +1,2 @@
+from pls_ import *
+from cca_ import *
